Building skeleton of website to put content into
Decided on two different fonts. One for navigation and another for text reading. I am also exploring the realm of customization with javascript which I have never done before so thats new for me. A good portion of the css and html is done for desktop viewing of the website. Figuring out how to do mobile will be interesting.