Finished the code necessary to make mobile possible
Added an onClick function to javascript to handle the nav menu being different on mobile and desktop. mobile will call a function to make the navigation menu appear and dissapear. Added a nav button that appears on mobile but not desktop. Css scales the website the way I want. Overall I am really happy with the results.