Fixed the overshoot problem and added menu closing on window resize
When going to about me or portfolio from the main homepage on mobile it would miss the section due to the viewport moving to a fixed position before the nav menu fully closes causing the website height to change but the viewport not changing with it. Fixed it by adding a link opening delay. Also added nav menu closing on resize. Might refactor the javascript at a later point.