Readded link highlighting function. Deleted extra portfolio containers
I realized that I accidentally deleted the function to highlight the links to indicate what section is currently being viewed. Readded that but also commented out the function call and will reenable it when there are enough portfolio projects to fully hide the about me tab. I may implement a percentage showing based system for the links and reenable it at that time.